Repo Pivot
This change pivots the repo from housing the test application
to housing a research and development prototype.  The intent
is to figure out what works for comparisions of SBOMs and extract
from this repo to another project in the near future.

This patch adds the data model for source and image SBOMS and
some of the organization around creating comparisons.

This patch also adds an enpoint `drift-debug` which will load
two given sbom files into the data model and open an ipython shell.

Signed-off-by: Ryan Brady <ryan.brady@anchore.com>